Create a folder and added two files, factList and FunFact component - updated index page
