+++ src/components/App/index.tsx
@@ -1,21 +1,15 @@
 import React from "react";
 import "./styles.css";
+import FunFacts from "../FunFacts/FunFact";
 
 const App: React.FC = () => {
   return (
     <div className="App">
       <header className="App-header">
         <p>
-          Edit <code>src/App.tsx</code> and save to reload.
+          Fun Facts
+          <FunFacts />
         </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
       </header>
     </div>
   );
