--- src/components/FunFact/factList.tsx
@@ -1,38 +1,0 @@
-interface Fact {
-    title: string,
-    image?: string,
-    id: number,
-    content: string,
-    category: string,
-}
-
-export const FunFacts: Fact[] = [
-    {
-        title: "Moons",
-        image: "http://something.com ",
-        id: 1,
-        content: "Mars has two moons ",
-        category: "rover",
-    },
-    {
-        title: "atmosphere",
-        image: "http://something.com ",
-        id: 2,
-        content: "Mars has two moons ",
-        category: "environment",
-    },
-    {
-        title: "max temperature",
-        image: "http://something.com ",
-        id: 3,
-        content: "Mars has two moons ",
-        category: "missions",
-    }
-    {
-        title: "Rover lands on mars",
-        image: "http://something.com ",
-        id: 4,
-        content: "Mars has two moons ",
-        category: "rover",
-    }
-]
