+++ src/components/FunFacts/FunFact.tsx
@@ -1,0 +1,10 @@
+import { factList } from "./factList";
+
+export default function FunFacts() {
+  return (
+    <div>
+      <p>factList</p>
+      <img src="https://picsum.photos/id/600/1600/900.jpg"></img>
+    </div>
+  );
+}
